fully schema-less buckets
No more 503 errors for "non-existent" buckets.  If the schema for a bucket has not been explicitly set, then it is assumed that any fields are allowed in the object, no fields are required, and all fields are both readable and writeable.

These "anything goes" settings are represented by "*" as the value of the allowed_fields, read_mask, or write_mask fields of a bucket schema.

It is now also not necessary to include all of allowed_fields, required_fields, read_mask, and write_mask in every schema PUT.  If any are left out, they are set to their defaults (allowed_methods="*", required_methods=[], read_mask="*", write_mask="*").

Just start storing data!