Allow streaming of keys over the raw interface.
Streaming is done using chunked-encoding where each chunk is a JSON
object containing a number of keys.  To invoke streaming, set the "keys"
query parameter to "stream". The first chunk is the bucket props unless you
set the "props" parameter to "false", in which case it will be an
empty JSON object.