Refactored riak kv backend modules to make handle_info work under the new vnode infrastructure.
Created a riak_kv_backend behavior and changed all the backend modules to use it.
Added a new callback/3 function for the modules and renamed the appropriate handle_info
calls.
Added riak_kv_backend:callback_after function to allow backends to schedule a callback
that will pass through the vnode and updated the vnode to pass on the callback message.
Made riak_kv_multi_backend walk through it's configuration correctly and pass the callbacks
through to each module.  It is up to each module to ignore callbacks that it did not request.