Fixed the Makefile so that the distdir build action functions properly.
The action now does what I believe it did prior to the move from mercurial to git. It adds support for "archiving" repositories that come from both git and mercurial during the dependency inclusion phase. It shows an error message if the user isn't at a tagged revision before trying to do the commit.

This will only work on a revision that is tagged _after_ this Makefile is committed because checking out an earlier version of the repo will run with an earlier version of the script.