Document the vnode_cache_entries tunable of riak_kv.
The number of LRU cache entries can be tuned in riak, but there is no
mention of this in the configuration file. By exposing the knob, users
can to some extent control the memory usage of riaks cache layer, saving
some memory. This is especially useful in development with low-memory
environments.