Set +x bit on all scripts in rel/files and update deb rules to ensure binaries are set +x
Debian packages were not having all executable scripts set with the executable bit.  Two things could fix this, the first being to properly set the scripts in the original reltool directory.   The second is to ensure they were getting set during the debian install, which was happening correctly in the rpm and solaris package managers.