Add staged clustering support to riak-admin
Update `riak-admin` to use the staged clustering support added in
response to basho/riak_core#177. The staged cluster interface is
provided by `riak-admin cluster` which provides commands to stage
a set of changes as well as the plan, commit, and clear commands
to view the set of pending changes and commit/clear them.