Add a pkgdep and origin line to the freebsd package to fix installation errors
Riak depends on openssl, but openssl is not installed by default on a base
FreeBSD 9 machine.  This adds a `@pkgdep` line to require openssl.

Fixes basho/riak#171

If an ORIGIN comment is not found, all new package installs and deletes
will comment on riak not having an origin recorded.

Fixes basho/riak#167