Create SMF service for EPMD on SmartOS
Riak is started and stopped on SmartOS using the [SMF system](http://wiki.joyent.com/display/smart/About+the+Service+Management+Facility). This system keeps track of every process started by a service.

This was causing difficulties with [epmd](http://www.erlang.org/doc/man/epmd.html) because that process would not shut down on `svcadm disable riak` causing Riak to go into a "maintenance" mode.

This pull request creates a separate SMF service for epmd called "riak-epmd".  This will do its best to start and stop epmd in a sane manner, allowing for Riak to be started and stopped also in sane way.

If an epmd process is active, but with no registered names, it is killed and restarted.

See: basho/riak_ee#/75