Backup config files and ensure proper ownership in Smartos
In some cases config files in `etc/riak` and `lib/riak` were
still not given the proper permissions on install. This sets
the ownership outside of the +CONTENTS file in the pre & post
install scripts. This addresses bug basho/riak#200

Also, a change to not overwrite existing config files (`app.config`
and `vm.args`) has been added on to the pre-install script.