Use erlang `os:getpid` as basis of `riak stop`
In BSD systems the regular expression was failing that attempted
to capture the PID of the running Riak node.  This PID was used
in the `riak stop` command to properly wait for a node to stop.

This change fixes the situation by replacing the fragile shell
methods of getting the PID with an erlang / nodetool based
method.  In addtion a `riak getpid` function has been added
for general admin use.

Addresses #251