Move SSL distribution items into a regular configuration file.
Before this change, `sed` was used to find some "magic comments" in
vm.args and slurp them into a file that is writeable by the `riak`
script, so that `nodetool` could use that file as arguments to its
init process, and thus be able to connect to a Riak node using the SSL
distribution.

This is wholly unnecessary, given that the emulator can accept
multiple -args_file command-line arguments. Instead, we now separate
the totally-optional-and-mostly-unused SSL settings into
etc/ssl_distribution.args, which can be used independently by
`nodetool` and in conjunction with vm.args for the normal node boot
process.

In order to support upgrades, the `riak` script checks for the
presence of the old sed-slurped file. If it is present and contains
active configuration items, a warning is presented and the script
exits non-zero. If the file exists and only contains comments, it is
deleted. Downgrades are not explicitly supported.