Move the stub 'riak' app into a subdir to avoid confusing rebar
Right now, if you try to do ./rebar eunit apps=riak_kb from the toplevel
of riak, rebar will just return. This is because it sees the ebin
directory at the toplevel and somehow it gets confused. The more
'release' friendly directory path is to have apps/riak/(ebin|src).

This branch makes that change and updates the rebar.config and
reltool.config files accordingly