Add `cuttlefish_conf` to pkg.vars.config to support packaging
The `cuttlefish_conf` variable was already in the vars.config
files, but that was not enough to support proper packaging
support of cuttlefish.  Many packages need to know about the
config files that are being installed on the system and the
vars.config template variables are not available to files outside
of the ones overlayed in the reltool.config.