Fix issues with scripts when CDPATH is set
The CDPATH environment variable can cause the cd command to sometimes
produce output. This breaks `runner_script_dir` detection in several
scripts. This fix suppresses the output from the cd command whenever it
is produced and allows the scripts to successfully detect the correct
directory for `runner_script_dir`.

For more information on when cd might produce output, see:
http://pubs.opengroup.org/onlinepubs/9699919799/utilities/cd.html#tag_20_14_10