Add "special" deps for make test
Some deps (riak_ensemble, meck, webmachine) don't work with the previous
`make test` command. riak_ensemble requires it's own script that runs
each test in it's own vm. webmachine has special test deps that it grabs
locally, and meck is meck (it has deps it expects to be in meck/deps at
test time.) All these could be addressed by a move to rebar3 I think (in
the future.) NOTE: there is a risk/problem around the versions of deps
in the riak_ensemble and webmachine deps and those shipped via
locked-deps. Care will have to be taken.