Remove riak-search, Makefile changes for test, and deps twiddling (#931)
* Remove vestiges of riak_search
* label edges of dep graph by version
* Add "special" deps for make test

Some deps (riak_ensemble, meck, webmachine) don't work with the previous
`make test` command. riak_ensemble requires it's own script that runs
each test in it's own vm. webmachine has special test deps that it grabs
locally, and meck is meck (it has deps it expects to be in meck/deps at
test time.) All these could be addressed by a move to rebar3 I think (in
the future.) NOTE: there is a risk/problem around the versions of deps
in the riak_ensemble and webmachine deps and those shipped via
locked-deps. Care will have to be taken.

* Add repl into release config.
* Add riak-repl script.
* Make rel/devrel use locked-deps add a lock file
* Hamcrest is also special
* Revert "label edges of dep graph by version"

This reverts commit 04b2e64e89ed1df1c198ccfb2edee710e01e36a0.  Bryan
already added and merged it, and this is simpler than resolving the
conflict

* Bump lock for merged 2.2.5->2.2