Update lock
Update lock -

riak_core (wit updated reference to basho eleveldb)
eleveldb (point at the basho repo not martinsumner)
yokozuna (taking in updated develop-2.9 with security fixes)
eper (this is the actuall referred to tag, this had previously been regressed in the rebar.config.lock because the tests din't appear to pass).

Change to the makefile so that eper tests are run using 'make test' - this was the problem with eper tests falsely reporting as failing.