Tidy make test
Testing all dependcies no works.  The riak-level eunit tests don't run as part of the test dependency run (as they fail due to dependancy issues - but pass as rebar3 eunit).

Lager does not run eunit tests (lager fails on OTP22 on the latest release - and as an external dependancy we have to live with that for now).  bahso_bench doesn't either (it has no eunit tests and is ony pulled in to handle problems with yokozuna testing).

The make rel now copies the release into rel/riak as per Riak 2.9 - for consistency.