Support debian make
Make profile for debian.  Issues exist when compared to 2.9 with permissions:

To help with an expectation that configuration will be written to usr/lib/riak - a system of links is used instead - with the links configured through the riakpre script.

Care needs to be taken with the merging of extended_start_script_hooks across profiles - this may not yield the expected results.  Hence each profile has a dedicated configurtaion and a single profile is used for make rel-deb.

The pid_dir is defined in multiple places (debian/vars.config, rebar.config, riak.riak.service).  These need to align.