script name is passed through from riak-admin to clique (#1025)
* script name is passed through from riak-admin to clique. It is currently set
to the base script, which is now `riak`. We changed some calls so that they
worked by suffixing the call with `-admin` to work around it.

There was an issue with running `riak admin handoff *`, which slipped through the
net. Looking at this again, seemed better to change the $SCRIPT var rather than
suffix the call, just in case there's anything else that we haven't caught, it
should account for it.

* add pre start hooks to dev profile in rebar.config and fix make devrel to use dev rebar profile.