Mas i1037 riakdiag (#1047)
Re-introduce capability to run `sudo riak debug` (not riak-debug).

This is considered as a special case, and bypasses the standard relx generated riak start script.  There were a number of issues with individual debug scripts related to permissions and the relative start location when running via he relx generated script.

This adds back into riak the riaknostic library, allowing for `riak admin diag` to be run.  The level of diagnostic checks supplied via this library are minimal at present though.  The diagnostic checks will be run by default as part of `riak debug` 

Co-authored-by: Bob-The-Marauder <nickadams_98@yahoo.com>