Temp replace reip with reip_manual (#1122)
* Temp replace reip with reip_manual

Reip was broken by the missing scripts necessary to generate the config so that erl could be run without the riak node active, but with the config necessary to load riak_core.

Now loading riak_core is not necessary with riak_manual .. but both the path to the ring directory, and the cluster name need to be manually passed in as command params.

* Update rebar.lock

* Update branch