Use tag for rebar3_cuttlefish
The rebar.lock file is not used for project_plugins, and so a tag must be used to consistently bring the same version of this software into the release.

The tagging of all dependencies has not been sone as part of this release.  This is to avoid confusion over tagging between 3.0 and 3.2 release paths, and remove some time-consuming activity prone to human error, form the release cycle.